Extracts ShortLinkShare
